refactor objective.Function out of Search
Will be useful to build interface toward external solvers.

- The objective.Function is now a member of Search, as `self.pb`.
- Some method were not used, I moved them based on their name.
- I tried to keep the code formatting style, but got rid of useless whitespaces anyway.
